+++ graph/src/main/java/com/social/graph/repository/ProfileRepository.java
@@ -17,11 +17,11 @@ public interface ProfileRepository extends Neo4jRepository<Profile, String> {
     List<Profile> findFriendsByProfileIdentity(@Param("identity") String identity);
 
     @Query(FIND_HOW_MANY_PROFILES_LIKE_THE_POST_TEMPLATE)
-    long findLikesAPostProfileCount(@Param("postIdentity") String postIdentity);
+    int findLikesAPostProfileCount(@Param("postIdentity") String postIdentity);
 
     @Query(FIND_HOW_MANY_PROFILES_DISLIKE_THE_POST_TEMPLATE)
-    long findDislikesAPostProfileCount(@Param("postIdentity") String postIdentity);
+    int findDislikesAPostProfileCount(@Param("postIdentity") String postIdentity);
 
     @Query(FIND_HOW_MANY_PROFILES_STAR_THE_POST_TEMPLATE)
-    long findStarsAPostProfileCount(String postIdentity);
+    int findStarsAPostProfileCount(String postIdentity);
 }
